requirements.txt
# annotated-types==0.6.0
# anyio==4.3.0
# asyncpg==0.29.0
# cffi==1.16.0
# click==8.1.7
# colorama==0.4.6
# cryptography==42.0.5
# ecdsa==0.19.0
# fastapi==0.110.3
# greenlet==3.0.3
# h11==0.14.0
# idna==3.7
# pyasn1==0.6.0
# pycparser==2.22
# pydantic==2.7.1
# pydantic_core==2.18.2
# python-jose==3.3.0
# rsa==4.9
# six==1.16.0
# sniffio==1.3.1
# SQLAlchemy==2.0.29
# starlette==0.37.2
# typing_extensions==4.11.0
# uvicorn==0.29.0



annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.0
fastapi==0.110.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
bcrypt==3.2.0
passlib==1.7.4
